# docs/swagger.yaml
basePath: /api
...
host: localhost:8080
info:
  contact:
    email: bitop@spatecon.ru
    name: API Support
    url: https://vk.com/bmstu_schedule
  description: API для работы с экзопланетами и звездами
  license:
    name: AS IS (NO WARRANTY)
  title: ExoCalc API
  version: "1.0"
paths:
...
  /calculate-exoplanets/remove-star/{id}:
    delete:
      consumes:
      - application/json
      description: Удаляет звезду из текущей заявки
      parameters:
      - description: ID звезды
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Удаление звезды из заявки
      tags:
      - CalculateExoplanets
...
  /selected-stars:
    get:
      consumes:
      - application/json
      description: Возвращает отфильтрованный список заявок
      parameters:
      - description: Дата начала (YYYY-MM-DD)
        in: query
        name: date_from
        type: string
      - description: Дата окончания (YYYY-MM-DD)
        in: query
        name: date_to
        type: string
      - description: Статус заявки (draft, formed, completed, declined)
        in: query
        name: status
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Получение списка заявок
      tags:
      - SelectedStars
    post:
      consumes:
      - application/json
      description: Создает новую черновую заявку
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Создание черновика заявки
      tags:
      - SelectedStars
  /selected-stars/{id}:
    delete:
      consumes:
      - application/json
      description: Удаляет заявку по ID
      parameters:
      - description: ID заявки
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Удаление заявки
      tags:
      - SelectedStars
    get:
      consumes:
      - application/json
      description: Возвращает заявку со всеми расчетами экзопланет
      parameters:
      - description: ID заявки
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Получение заявки по ID
      tags:
      - SelectedStars
    put:
      consumes:
      - application/json
      description: Обновляет дату и ученого для заявки
      parameters:
      - description: ID заявки
        in: path
        name: id
        required: true
        type: integer
      - description: Дата и ученый
        in: body
        name: input
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Обновление заявки
      tags:
      - SelectedStars
  /selected-stars/{id}/form:
    put:
      consumes:
      - application/json
      description: Переводит заявку из черновика в статус "formed"
      parameters:
      - description: ID заявки
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Формирование заявки
      tags:
      - SelectedStars
  /selected-stars/{id}/moderate:
    put:
      consumes:
      - application/json
      description: Модератор подтверждает или отклоняет заявку (требуется роль Astronomer)
      parameters:
      - description: ID заявки
        in: path
        name: id
        required: true
        type: integer
      - description: Действие (complete/decline) и ID модератора
        in: body
        name: input
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Модерация заявки
      tags:
      - SelectedStars
  /selected-stars/add-star/{id}:
    post:
      consumes:
      - application/json
      description: Добавляет звезду по ID в текущую заявку
      parameters:
      - description: ID звезды
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "201":
          description: Created
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Добавление звезды в заявку
      tags:
      - SelectedStars
  /selected-stars/count:
    get:
      consumes:
      - application/json
      description: Возвращает ID текущей черновой заявки и количество звезд в ней
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Получение количества звезд в текущей заявке
      tags:
      - SelectedStars
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: 'Введите токен в формате: Bearer {token}'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"